+++ lib/packages/core/src/api/factories/contract/index.ts
@@ -1,0 +1,4 @@
+/** @module core.api.factories */
+
+export * from './getContract';
+export * from './getContractsByAccount';