+++ lib/packages/core/src/typings/contractList.ts
@@ -1,0 +1,12 @@
+/** @module core */
+
+import {Contract} from './contract';
+
+/**
+ * Copyright (c) 2019 Burst Apps Team
+ */
+
+export interface ContractList {
+    ats: Contract[];
+    requestProcessingTime: number;
+}