+++ web/angular-wallet/src/app/main/transactions/transactions.resolver.ts
@@ -1,9 +1,8 @@
-import { Injectable } from '@angular/core';
+import {Injectable} from '@angular/core';
 
-import { Resolve, ActivatedRouteSnapshot } from '@angular/router';
-import { TransactionService } from './transaction.service';
-import { Transaction } from '@burstjs/core';
-import { AccountService } from 'app/setup/account/account.service';
+import {Resolve, ActivatedRouteSnapshot} from '@angular/router';
+import {Transaction} from '@burstjs/core';
+import {AccountService} from 'app/setup/account/account.service';
 
 @Injectable({
   providedIn: 'root'
@@ -12,7 +11,7 @@ export class TransactionsResolver implements Resolve<Promise<Transaction[]>> {
   constructor(private accountService: AccountService) {
   }
 
-  async resolve(route: ActivatedRouteSnapshot) {
+  async resolve(route: ActivatedRouteSnapshot): Promise<Transaction[]> {
     const account = route.params.id || this.accountService.currentAccount.getValue().account;
     try {
       const unconfirmedTransactions = await this.accountService.getUnconfirmedTransactions(account);