+++ web/angular-wallet/src/environments/environment.web.ts
@@ -1,0 +1,12 @@
+import {version} from '../../package.json';
+
+export const environment = {
+  production: true,
+  hmr: false,
+  version,
+  defaultNode: '',
+  market: {
+    tickerUrl: 'https://api.coinmarketcap.com/v1/ticker/burst/',
+    tickerInterval: 30 * 1000
+  }
+};